topic_scorer/requirements.txt
asyncpg>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.1
PyYAML>=6.0.1
aiohttp>=3.9.0
prometheus-client>=0.20.0
